proxy_server/requirements.txt
aiohttp==3.9.3
aiohttp-sse==2.1.0
python-dotenv==1.0.1
cryptography==42.0.5